services/browser-engine/app/scraper/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
scrapy==2.11.0
beautifulsoup4==4.12.2
trafilatura==1.6.4
requests==2.31.0
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-timeout==2.4.0
aiohttp==3.9.1
PyPDF2==3.0.1
websockets==12.0
httpx==0.25.2
pydantic-settings==2.0.3
